Separate background.js and some other things into separate modules Make sure modules all export the same symbol as their filename Add a little appropriate documentation
