+++ webextension/background/README.md
@@ -1,0 +1,7 @@
+These are all the files/modules for the background worker.  This is the higher-privileged WebExtension worker that has no interface.
+
+- `analytics.js` handles sending events to the server for use with GA
+- `auth.js` handles the initial authentication call, and handles subsequent requests using that authentication.  It also informs `analytics.js` of any A/B tests so those can be submitted.
+- `communication.js` handles communication with the content workers.
+- `deviceinfo.js` gets information about this device that is used in some logging and analytics.
+- `main.js` loads and launches everything.
