+++ webextension/background/analytics.js
@@ -1,0 +1,26 @@
+/* globals main */
+
+window.analytics = (function () {
+  let exports = {};
+
+  exports.sendEvent = function (action, label) {
+    let eventCategory = "addon";
+    let url = main.getBackend() + "/event";
+    let req = new XMLHttpRequest();
+    req.open("POST", url);
+    req.setRequestHeader("content-type", "application/json");
+    req.onload = () => {
+      if (req.status >= 300) {
+        console.warn("Event gave non-2xx response:", req.status);
+      }
+    };
+    // FIXME: add cdX and other details from req.js
+    req.send(JSON.stringify({
+      event: eventCategory,
+      action,
+      label
+    }));
+  };
+
+  return exports;
+})();
