+++ webextension/randomString.js
@@ -1,7 +1,6 @@
-/** Similar to addon/lib/randomstring, but uses Math.random */
 /* exported randomString */
 
-function randomString(length, chars) { // eslint-disable-line no-unused-vars
+window.randomString = function randomString(length, chars) {
   let randomStringChars = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789";
   chars = chars || randomStringChars;
   let result = "";
@@ -9,4 +8,4 @@ function randomString(length, chars) { // eslint-disable-line no-unused-vars
     result += chars.charAt(Math.floor(Math.random() * chars.length));
   }
   return result;
-}
+};
