+++ webextension/selector/shooter-interactive-worker.js
