+++ webextension/selector/snapping.js
