+++ webextension/selector/util.js
